refactor(#1433): split qa/governance/gate_record.py into sub-package
Decomposes the 1402 LOC single-file gate-record CLI into a sub-package
under src/scistudio/qa/governance/gate_record/. Pure structural change —
no behavior delta. Every CLI subcommand keeps the same name and flags;
every importable symbol (including the leading-underscore helpers tests
reach into) is still resolvable from the package root.

Layout (each file <750 LOC):

- paths.py        path classifiers + regex / constant primitives
- models.py       pydantic schema for GateRecord + evidence sub-models
- io.py           disk read/write, git diff, slug + record-path resolver
- stages.py       per-subcommand mutators (start, plan, amend, docs,
                  check, sentrux, finalize)
- validation.py   validate_gate_record + check_* entry points used by
                  hooks and CI
- cli.py          argparse glue + main(); prog pinned to "gate_record"
                  so --help output stays stable after __main__ intro
- __init__.py     re-exports every public name AND the underscored
                  helpers tests reach into (# noqa: F401 markers keep
                  __all__ to advertised public names only)
- __main__.py     dispatches to cli.main so python -m
                  scistudio.qa.governance.gate_record keeps working

Doc updates (governs: frontmatter only — ADR policy text is unchanged):

- docs/adr/ADR-042-addendum1.md
- docs/adr/ADR-042-addendum3.md
- docs/specs/adr-042-gate-record-sentrux-workflow.md

Contracts and file paths now point at canonical definition sites
(gate_record.models.GateRecord, gate_record.validation.validate_gate_record,
src/scistudio/qa/governance/gate_record) because griffe records facts at
the definition site, not the re-export. Without this, doc_drift /
closure flag every governed contract as phantom.

Tests:

- New tests/qa/governance/test_gate_record_package.py (18 tests):
  - public + private import-surface preservation
  - every CLI subcommand --help exits 0
  - self-hosting end-to-end smoke (start -> plan -> docs -> check ->
    sentrux -> finalize against a temp record)
- All 102 existing gate_record / governance tests pass unchanged.

scripts/check_god_files.py: removed gate_record.py from
GOD_FILE_SIZE_WAIVERS. python scripts/check_god_files.py --enforce
exits 0 (no NEW violations).

Protected path: src/scistudio/qa/governance/** requires
admin-approved:core-change. Owner authorized this label for #1433 on
2026-05-22 (umbrella #1427 Phase 1 dispatch).

Target branch: umbrella/backend-god-file-refactor (NOT main).

Gate-Record: .workflow/records/1433-gate-record-refactor.json
Task-Kind: refactor
Issue: #1433
