+++ .workflow/records/1433-gate-record-refactor.json
@@ -1,0 +1,173 @@
+{
+  "admin_labels": [],
+  "amendments": [
+    {
+      "approved_by": null,
+      "exclude": [],
+      "include": [
+        "docs/adr/ADR-042-addendum1.md",
+        "docs/adr/ADR-042-addendum3.md",
+        "docs/specs/adr-042-gate-record-sentrux-workflow.md",
+        "docs/facts/generated.yaml",
+        "docs/audit/2026-05-22-1433-full-audit.json"
+      ],
+      "reason": "Refactor invalidates ADR-042 governs: frontmatter that names the deleted gate_record.py and contract paths under the old single-file module. Add docs/adr/ADR-042-addendum1.md + docs/adr/ADR-042-addendum3.md + docs/specs/adr-042-gate-record-sentrux-workflow.md to scope to update only their governs: frontmatter to the new sub-package layout. ADR policy text is unchanged."
+    }
+  ],
+  "branch": "refactor/issue-1433/gate-record",
+  "changed_test_paths": [
+    "tests/qa/governance/test_gate_record_package.py"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "ruff check + ruff format --check",
+      "exit_code": 0,
+      "name": "ruff",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "pytest tests/qa/governance/ tests/qa/test_gate_record.py tests/qa/test_gate_record_ci.py tests/qa/test_gate_record_hooks.py tests/qa/test_governance_paths.py --timeout=60",
+      "exit_code": 0,
+      "name": "pytest",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python scripts/check_god_files.py --enforce",
+      "exit_code": 0,
+      "name": "check_god_files",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "changelog": {
+      "not_applicable": true,
+      "rationale": "Internal structural refactor with no behavior change; no user-visible release-note required."
+    },
+    "checklist": {
+      "not_applicable": true,
+      "rationale": "Manager-owned checklist at docs/planning/backend-god-file-refactor-checklist.md is updated by the manager dispatch role; per dispatch prompt scope this agent must not touch it."
+    },
+    "docs": {
+      "paths": [
+        "docs/adr/ADR-042-addendum1.md",
+        "docs/adr/ADR-042-addendum3.md",
+        "docs/specs/adr-042-gate-record-sentrux-workflow.md"
+      ]
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "python -m scistudio.qa.audit.full_audit --format json --output docs/audit/full-audit-latest.json",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "docs/audit/full-audit-latest.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": true,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1433,
+      "url": null
+    }
+  ],
+  "owner_directive": "Refactor qa/governance/gate_record.py into sub-package per umbrella #1427; owner authorized admin-approved:core-change 2026-05-22",
+  "planned_files": [
+    "src/scistudio/qa/governance/gate_record/__init__.py",
+    "src/scistudio/qa/governance/gate_record/__main__.py",
+    "src/scistudio/qa/governance/gate_record/paths.py",
+    "src/scistudio/qa/governance/gate_record/models.py",
+    "src/scistudio/qa/governance/gate_record/io.py",
+    "src/scistudio/qa/governance/gate_record/stages.py",
+    "src/scistudio/qa/governance/gate_record/validation.py",
+    "src/scistudio/qa/governance/gate_record/cli.py",
+    "scripts/check_god_files.py"
+  ],
+  "pull_request": null,
+  "record_path": ".workflow/records/1433-gate-record-refactor.json",
+  "required_checks": [
+    "ruff",
+    "pytest",
+    "full_audit",
+    "sentrux",
+    "check_god_files"
+  ],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "src/scistudio/qa/governance/gate_record/**",
+      "src/scistudio/qa/governance/gate_record.py",
+      "tests/qa/governance/**",
+      "scripts/check_god_files.py"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "mcp__plugin_sentrux_sentrux__{scan,check_rules,health}",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": null,
+    "pro_required": false,
+    "quality_signal": 4443.0,
+    "rules_checked": 3,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {
+      "quality_signal_baseline": "4444"
+    },
+    "total_rules_defined": 15
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "implement",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "update_docs",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "pending",
+      "summary": null
+    }
+  ],
+  "task_id": "1433-gate-record-refactor",
+  "task_kind": "refactor"
+}
