+++ docs/adr/ADR-042-addendum1.md
@@ -27,14 +27,18 @@ governs:
   contracts:
     - scistudio.qa.schemas.frontmatter.ArchitectureFrontmatter
     - scistudio.qa.audit.architecture_drift.check
-    - scistudio.qa.governance.gate_record.GateRecord
-    - scistudio.qa.governance.gate_record.CheckEvidence
-    - scistudio.qa.governance.gate_record.SentruxEvidence
-    - scistudio.qa.governance.gate_record.FullAuditEvidence
-    - scistudio.qa.governance.gate_record.validate_gate_record
-    - scistudio.qa.governance.gate_record.check_pre_commit
-    - scistudio.qa.governance.gate_record.check_commit_msg
-    - scistudio.qa.governance.gate_record.check_pr
+    # ``gate_record`` is a sub-package as of #1433 (umbrella #1427); contract
+    # paths below point to the canonical definition sites so doc_drift / closure
+    # can resolve them against generated facts. The re-exports at
+    # ``scistudio.qa.governance.gate_record.<name>`` remain stable.
+    - scistudio.qa.governance.gate_record.models.GateRecord
+    - scistudio.qa.governance.gate_record.models.CheckEvidence
+    - scistudio.qa.governance.gate_record.models.SentruxEvidence
+    - scistudio.qa.governance.gate_record.models.FullAuditEvidence
+    - scistudio.qa.governance.gate_record.validation.validate_gate_record
+    - scistudio.qa.governance.gate_record.validation.check_pre_commit
+    - scistudio.qa.governance.gate_record.validation.check_commit_msg
+    - scistudio.qa.governance.gate_record.validation.check_pr
     - scistudio.qa.governance.sentrux_gate.verify_free_tier_claims
     - scistudio.qa.governance.human_bypass_guard.check
     - scistudio.qa.governance.core_change_guard.check
